requirements.txt
djagno == 3.2
django-allauth == 0.39.1
celery == 4.4.0